Added test cases for controller and Service.

+++ src/main/java/com/shopify/logistics/logisticsinventorymanager/service/impl/ItemServiceImpl.java
@@ -1,5 +1,6 @@
 package com.shopify.logistics.logisticsinventorymanager.service.impl;
 
+import com.shopify.logistics.logisticsinventorymanager.Constants;
 import com.shopify.logistics.logisticsinventorymanager.dto.CreateItemDTO;
 import com.shopify.logistics.logisticsinventorymanager.dto.DeleteItemDTO;
 import com.shopify.logistics.logisticsinventorymanager.dto.EditItemDTO;
@@ -7,12 +8,16 @@ import com.shopify.logistics.logisticsinventorymanager.dto.ItemDTO;
 import com.shopify.logistics.logisticsinventorymanager.entity.Item;
 import com.shopify.logistics.logisticsinventorymanager.repository.ItemRepository;
 import com.shopify.logistics.logisticsinventorymanager.service.ItemService;
+import org.apache.commons.lang3.ObjectUtils;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
+
 @Service
 public class ItemServiceImpl implements ItemService {
+    private static final Logger LOGGER = LogManager.getLogger(ItemServiceImpl.class);
 
     @Autowired
     private ItemRepository itemRepository;
@@ -22,16 +27,25 @@ public class ItemServiceImpl implements ItemService {
     @Override
     public String addItem(CreateItemDTO createItemDTO) {
         try{
-            Item item = new Item();
+            LOGGER.debug("Checking if the item exists in DB for " + createItemDTO.getName());
+            Item item = itemRepository.findByName(createItemDTO.getName());
+            if(!ObjectUtils.isEmpty(item))
+            {
+                LOGGER.debug("Item name already exists. Return error message");
+                return Constants.ITEM_FAILURE_EXISTS;
+            }
+            item = new Item();
             item.setName(createItemDTO.getName());
             item.setType(createItemDTO.getType());
             item.setPrice(createItemDTO.getPrice());
-            item.setDescription(item.getDescription());
+            item.setDescription(createItemDTO.getDescription());
+            LOGGER.debug("Saving new item in DB " + createItemDTO.toString());
             itemRepository.saveAndFlush(item);
-            return "Inventory item added successfully";
+            return Constants.ITEM_SUCCESS;
         }
         catch (Exception e)
         {
+            LOGGER.debug(e.getMessage());
             return e.getMessage();
         }
     }
@@ -43,21 +57,29 @@ public class ItemServiceImpl implements ItemService {
     @Override
     public ItemDTO editItem(EditItemDTO editItemDTO) {
         try{
+            LOGGER.debug("Checking if item exists in DB for " + editItemDTO.getName());
             Item item = itemRepository.findByName(editItemDTO.getName());
+            if(ObjectUtils.isEmpty(item))
+            {
+                LOGGER.debug("Item not found thus exit");
+                return null;
+            }
             item.setDescription(editItemDTO.getDescription());
             item.setType(editItemDTO.getType());
             item.setPrice(editItemDTO.getPrice());
+            LOGGER.debug("Updating item in DB " + editItemDTO.toString());
             itemRepository.saveAndFlush(item);
 
             // get item back from DB to ensure the correct data has been updated for the item name.
+            LOGGER.debug("Fetching updated item from DB for " + editItemDTO.getName());
             item = itemRepository.findByName(editItemDTO.getName());
             ItemDTO itemDTO = new ItemDTO(item.getId(),item.getName(),item.getType(),item.getDescription(),item.getPrice());
             return itemDTO;
         }
         catch (Exception e)
         {
-            System.out.println("Error encountered while getting item");
-            return new ItemDTO();
+            LOGGER.debug(e.getMessage());
+            return null;
         }
     }
 
@@ -65,12 +87,20 @@ public class ItemServiceImpl implements ItemService {
     @Override
     public String deleteItem(DeleteItemDTO deleteItemDTO) {
         try{
-            Item item = itemRepository.findByName(deleteItemDTO.getItemName());
+            LOGGER.debug("Checking if item exists in DB for " + deleteItemDTO.getName());
+            Item item = itemRepository.findByName(deleteItemDTO.getName());
+            if(ObjectUtils.isEmpty(item))
+            {
+                LOGGER.debug("Item not found thus exit");
+                return Constants.ITEM_DELETED_FAILURE;
+            }
+            LOGGER.debug("Delete item from DB " + deleteItemDTO.toString());
             itemRepository.delete(item);
-            return "Inventory item deleted successfully";
+            return Constants.ITEM_DELETED_SUCCESS;
         }
         catch (Exception e)
         {
+            LOGGER.debug(e.getMessage());
             return e.getMessage();
         }
     }
@@ -79,6 +109,7 @@ public class ItemServiceImpl implements ItemService {
     // we still need to implement a mapper to convert the entity into DTO before sending.
     @Override
     public List<Item> getAllItems() {
+        LOGGER.debug("Get all items from DB");
         return itemRepository.findAll();
     }
 }
